marsadm: circumvent deprecated _get_actual_primary()
This should vanish definitely, it is broken by concept.

It cannot work. In split brain situations, there are many
actual primaries (by definition).

The only place where it seemingly was needed was check_primary_gone(),
and even there it was not really needed.
Now we simply check that the number of actual primaries is 0.