doc: update user manual appendix B
Damn. This was not updated for years.

There are much better incident resolution methods in the meantime.

Unfortunately, the old resolution process was prescribed by somebody
else, who is no longer in the company. And I was commanded to
transfer a very old internal handout onto github.

I was NOT the author of the old internal handout. Unfortunately,
anyone not knowing this history but reading now at github can draw the
wrong conclusion that the so-called "handout" in the LyX format
were fully authored by me.

Fortunately, I have much more freedom now. It needs to be
called "guide" in place of "handout". I am not responsible
for the setup of any particular instance of the MARS
component as configured and operated by any team
anywhere over the world.

To avoid any potential breakage of operational processes I have not
under control, I do not change the highlevel section structure.

I just update the concrete commands, correct some potential
misunderstandings, and add some more hints and pitfall warnings.

I am happy to receive constructive feedback on this.