all: remove historic workaround
See 495490f4b6d2bc2c5efe2ff6ea41eadb471be448

The kernel does not really need this, only
in very special situations, and only when the old
ssh-based method is working as a substotite of the
modern push operations.

Essentially, this was a workaround for a workaround.

Remove this for reduction of unnecessary complexity.