main: fix recovery upon defective logfile without successor
Workaround was possible by manually providing an empty next logfile, forcing
an internal rotate operation. However, IMHO this should not be exposed to
userspace, but functionally migrated to kernelspace in the long term.

Otherwise, possible races between userspace and kernelspace may explode.
Control over logfiles content is not enough; even the sheer existence
should be the responsibility of kernel code (as much as possible).

Future fixes and improvements should go in this direction.
Thus I don't add an extra marsadm command for creation of empty
logfiles.

The current marsadm code in cron (and siblings) should be decommissioned
some day, after the kernel is fully responsible, even for create-resource
and so on.