server: earlier disallow IO ops on unconnected bricks
Unconnected server bricks cannot work at the IO layer, but
are required for the strategy layer (in order to get strategic
information over the network even when IO is not yet possible).

Reflect this in the code, and report any potential communication
mischief as early as possible (safeguard).

Functional testing did not show a difference, but timing
behaviour should have improved.