net: fix spelling of constants
The typo in mars_recv_raw() was equivalent to MSG_CONFIRM, which should
not imply a functional harm, but obviously any nonblocking
reads on sockets did not work as intended (by staying
a synchronous operation).

Two other potential typos now corrected, without causing a difference
in the resulting constant values.

Side note 1: kernel_connect() is documented in net/socket.c that the
spelling O_NONBLOCK is the right one.

Side note 2: kernel_accept() is documented in net/socket.c that the
spelling SOCK_NONBLOCK is the right one for this case.
However, this in turn is defined as equivalent to O_NONBLOCK
in include/linux/net.h (thus no real difference).

Overall: humans may be misleaded.